fix: reuseable fixtures removed from repo_tool utils
