--- dashboard_dependencies_ymls/datafiles/super-csv_dependencies_health.yaml
@@ -1,99 +1,0 @@
-TIMESTAMP: 2023-12-01
-dependencies:
-    count: 123
-    github:
-        count: 0
-        list: '[]'
-    js:
-        count: 0
-        list: ''
-    js.all:
-        count: 0
-        list: ''
-    js.dev:
-        count: 0
-        list: ''
-    pypi:
-        count: 26
-        list: '["kombu==4.6.11", "djangorestframework==3.14.0", "pbr==5.11.1", "typing-extensions==4.7.1",
-            "edx-django-utils==5.6.0", "pytz==2023.3", "stevedore==5.1.0", "django-crum==0.7.9",
-            "asgiref==3.7.2", "pynacl==1.5.0", "django==3.2.20", "django-model-utils==4.3.1",
-            "sqlparse==0.4.4", "vine==1.3.0", "click==8.1.6", "newrelic==8.9.0", "psutil==5.9.5",
-            "cffi==1.15.1", "amqp==2.6.1", "celery==4.4.7", "billiard==3.6.4.0", "edx-celeryutils==1.2.3",
-            "pycparser==2.21", "django-waffle==4.0.0", "simplejson==3.19.1", "jsonfield==3.1.0"]'
-    pypi_all:
-        count: 123
-        list: '["pytest-django==4.5.2", "pyyaml==6.0.1", "twine==4.0.2", "importlib-resources==6.0.0",
-            "iniconfig==2.0.0", "stevedore==5.1.0", "django-crum==0.7.9", "coverage[toml]==7.2.7",
-            "pylint-django==2.5.3", "pyproject-hooks==1.0.0", "newrelic==8.9.0", "click==8.1.3",
-            "distlib==0.3.7", "amqp==2.6.1", "more-itertools==10.0.0", "billiard==3.6.4.0",
-            "pydata-sphinx-theme==0.13.3", "imagesize==1.4.1", "accessible-pygments==0.0.4",
-            "jaraco-classes==3.3.0", "py==1.11.0", "edx-celeryutils==1.2.3", "jsonfield==3.1.0",
-            "code-annotations==1.5.0", "sqlalchemy==2.0.19", "rfc3986==2.0.0", "pylint==2.17.5",
-            "requests-toolbelt==1.0.0", "importlib-metadata==6.8.0", "certifi==2023.7.22",
-            "six==1.16.0", "tomlkit==0.12.1", "edx-i18n-tools==1.0.0", "rich==13.5.1",
-            "django-model-utils==4.3.1", "readme-renderer==40.0", "sqlparse==0.4.4",
-            "pygments==2.15.1", "keyring==24.2.0", "celery==4.4.7", "chardet==5.1.0",
-            "idna==3.4", "exceptiongroup==1.1.2", "pip-tools==7.1.0", "mdurl==0.1.2",
-            "bleach==6.0.0", "sphinx==6.2.1", "filelock==3.12.2", "soupsieve==2.4.1",
-            "alabaster==0.7.13", "pkginfo==1.9.6", "virtualenv==20.24.2", "edx-lint==5.3.4",
-            "diff-cover==7.7.0", "pbr==5.11.1", "typing-extensions==4.7.1", "edx-django-utils==5.6.0",
-            "platformdirs==3.10.0", "greenlet==2.0.2", "wrapt==1.15.0", "freezegun==1.2.2",
-            "babel==2.12.1", "astroid==2.15.6", "pynacl==1.5.0", "sphinxcontrib-htmlhelp==2.0.1",
-            "wheel==0.41.0", "tomli==2.0.1", "tox-battery==0.6.1", "charset-normalizer==3.2.0",
-            "sphinxcontrib-devhelp==1.0.2", "vine==1.3.0", "build==0.10.0", "webencodings==0.5.1",
-            "markupsafe==2.1.3", "pytest==7.4.0", "psutil==5.9.5", "cffi==1.15.1",
-            "urllib3==2.0.4", "sphinxcontrib-jsmath==1.0.1", "jinja2==3.1.2", "pycparser==2.21",
-            "django-waffle==4.0.0", "packaging==23.1", "sphinxcontrib-applehelp==1.0.4",
-            "mccabe==0.7.0", "sphinxcontrib-qthelp==1.0.3", "restructuredtext-lint==1.4.0",
-            "kombu==4.6.11", "docutils==0.19", "djangorestframework==3.14.0", "pytest-cov==4.1.0",
-            "dill==0.3.7", "mock==5.1.0", "beautifulsoup4==4.12.2", "lazy-object-proxy==1.9.0",
-            "pytz==2023.3", "zipp==3.16.2", "python-dateutil==2.8.2", "pylint-plugin-utils==0.8.2",
-            "requests==2.31.0", "asgiref==3.7.2", "pylint-celery==0.3", "tox==3.28.0",
-            "sphinx-book-theme==1.0.1", "polib==1.2.0", "pip==23.2.1", "sphinxcontrib-serializinghtml==1.1.5",
-            "django==3.2.20", "pluggy==1.2.0", "path==16.7.1", "click-log==0.4.0",
-            "click==8.1.6", "python-slugify==8.0.1", "snowballstemmer==2.2.0", "ddt==1.6.0",
-            "isort==5.12.0", "setuptools==68.0.0", "markdown-it-py==3.0.0", "text-unidecode==1.3",
-            "doc8==1.1.1", "pydocstyle==6.3.0", "pycodestyle==2.11.0", "simplejson==3.19.1"]'
-django:
-    '4.0': v3.1.0
-    '4.1': null
-    '4.2': null
-    has_django: true
-django_packages:
-    django_32:
-        count: 12
-        list: '["code-annotations", "edx-celeryutils", "djangorestframework", "pytest-django",
-            "newrelic", "django-waffle", "django-model-utils", "django-crum", "pylint-django",
-            "jsonfield", "edx-django-utils", "edx-i18n-tools"]'
-    total:
-        count: 12
-        list: '["code-annotations", "edx-celeryutils", "djangorestframework", "pytest-django",
-            "newrelic", "django-waffle", "django-model-utils", "django-crum", "pylint-django",
-            "jsonfield", "edx-django-utils", "edx-i18n-tools"]'
-    upgraded:
-        count: 12
-        list: '["code-annotations", "edx-celeryutils", "djangorestframework", "pytest-django",
-            "newrelic", "django-waffle", "django-model-utils", "django-crum", "pylint-django",
-            "jsonfield", "edx-django-utils", "edx-i18n-tools"]'
-github:
-    branch_count: 12
-    org_name: openedx
-    pulls_count: 8
-python:
-    '3.10': null
-    '3.11': null
-    '3.8': 0.9.8
-    '3.9': null
-requires:
-    boto: false
-    django: true
-    nose: false
-    pytest: true
-setup_py:
-    py38_classifiers: true
-    pypi_name: super-csv
-    python_versions:
-    - '3'
-    - '3.8'
-    repo_url: https://github.com/openedx/super-csv
