+++ dependencies_health/check_django_dependencies_compatibility.py
@@ -1,23 +1,26 @@
 """
- contains check that reads/parses dependencies of a repo
+Contains check that reads/parses dependencies of a repo
 """
 
 import logging
 
-from repo_health.utils import set_django_packages
+from repo_health.check_django_dependencies_compatibility import (
+    set_django_packages,
+    django_dependency_sheet_fixture
+)
 
 logger = logging.getLogger(__name__)
 
 MODULE_DICT_KEY = "django_packages"
 
 
-def check_django_dependencies_status(repo_path, all_results, django_deps_sheet):
+def check_django_dependencies_status(repo_path, all_results, django_dependency_sheet_fixture):
     """
     Test to find the django dependencies compatibility
     """
     all_results = set_django_packages(
         repo_path,
         all_results,
-        django_deps_sheet,
+        django_dependency_sheet_fixture,
         MODULE_DICT_KEY
     )
